Process Count, Test Generator
Enabled user to define the number of tests to be generated, and a range of process counts for them instead of a hard value.

TODO: Enable the user to run multiple tests in a row and display only averages instead of the whole output.

Included one such set of 100 tests generated through this method.