Batch Testing + Batch Result Output Complete
Added batched testing as well as result output and formatting as .csv's. Tests are also now generated as .json, while this doesn't make much of a functional difference from .txt, it's more correct.

Results are stored in the directory the batch is taken from, and the algorithm in use for that run is appended to the results name. Re-doing a test with the same algorithm will overwrite the previous file, but this should result in no changes anyway.